abstracted modes, improved perf and added tests
