+++ cypress/integration/circlemarker.spec.js
@@ -46,7 +46,6 @@ describe('Draw Circle Marker', () => {
 
 
   it('handles 6k circle markers in under 1 sec', () => {
-    let t0;
 
     cy.toolbarButton('circle-marker')
       .click()
@@ -54,53 +53,7 @@ describe('Draw Circle Marker', () => {
     cy.get(mapSelector)
       .click(150, 250)
 
-    cy.window().then(({ map, L }) => {
-      t0 = performance.now();
-
-      function getRandomLatLng() {
-        const bounds = map.getBounds()
-        const southWest = bounds.getSouthWest()
-        const northEast = bounds.getNorthEast()
-        const lngSpan = northEast.lng - southWest.lng
-        const latSpan = northEast.lat - southWest.lat;
-
-        return new L.LatLng(
-          southWest.lat + latSpan * Math.random(),
-          southWest.lng + lngSpan * Math.random());
-      }
-
-      const terminals = [];
-      const locations = [];
-
-      for (let i = 0; i < 3500; i += 1) {
-        locations.push(L.circleMarker(getRandomLatLng(map)));
-
-      }
-
-      for (let i = 0; i < 2500; i += 1) {
-        terminals.push(L.circleMarker(getRandomLatLng(map)));
-      }
-
-      const t = L.layerGroup(terminals).addTo(map)
-      const l = L.layerGroup(locations).addTo(map)
-
-      const base = {}
-
-      const overlays = {
-        "Locations": t,
-        "Terminals": l
-      }
-
-      L.control.layers(base, overlays).addTo(map)
-    });
-
-    cy.window().then(() => {
-      const t1 = performance.now();
-      const delta = Math.abs(t1 - t0);
-      console.log(`Rendering 6k CircleMarkers took ${delta} milliseconds.`);
-
-      expect(delta).to.lessThan(1000);
-    });
+    cy.testLayerAdditionPerformance();
   });
