+++ src/js/L.PM.Map.js
@@ -4,11 +4,13 @@ import translations from '../assets/translations';
 import Utils from './L.PM.Utils'
 
 import GlobalEditMode from './Mixins/Modes/Mode.Edit';
+import GlobalDragMode from './Mixins/Modes/Mode.Drag';
+import GlobalRemovalMode from './Mixins/Modes/Mode.Removal';
 
 const { findLayers } = Utils
 
 const Map = L.Class.extend({
-  includes: [GlobalEditMode],
+  includes: [GlobalEditMode, GlobalDragMode, GlobalRemovalMode],
   initialize(map) {
     this.map = map;
     this.Draw = new L.PM.Draw(map);
@@ -59,149 +61,7 @@ const Map = L.Class.extend({
   setPathOptions(options) {
     this.Draw.setPathOptions(options);
   },
-  removeLayer(e) {
 
-    const layer = e.target;
-    // only remove layer, if it's handled by leaflet-geoman,
-    // not a tempLayer and not currently being dragged
-    const removeable =
-      !layer._pmTempLayer && (!layer.pm || !layer.pm.dragging());
-
-    if (removeable) {
-      layer.remove();
-      this.map.fire('pm:remove', { layer });
-    }
-  },
-  globalDragModeEnabled() {
-    return !!this._globalDragMode;
-  },
-  enableGlobalDragMode() {
-    const layers = findLayers(this.map);
-
-    this._globalDragMode = true;
-
-    layers.forEach(layer => {
-      layer.pm.enableLayerDrag();
-    });
-
-    // remove map handler
-    this.map.on('layeradd', this.layerAddHandler, this);
-
-    // toogle the button in the toolbar if this is called programatically
-    this.Toolbar.toggleButton('dragMode', this._globalDragMode);
-
-    this._fireDragModeEvent(true);
-  },
-  disableGlobalDragMode() {
-    const layers = findLayers(this.map);
-
-    this._globalDragMode = false;
-
-    layers.forEach(layer => {
-      layer.pm.disableLayerDrag();
-    });
-
-    // remove map handler
-    this.map.off('layeradd', this.layerAddHandler, this);
-
-    // toogle the button in the toolbar if this is called programatically
-    this.Toolbar.toggleButton('dragMode', this._globalDragMode);
-
-    this._fireDragModeEvent(false);
-  },
-  _fireDragModeEvent(enabled) {
-    this.map.fire('pm:globaldragmodetoggled', {
-      enabled,
-      map: this.map,
-    });
-  },
-  toggleGlobalDragMode() {
-    if (this.globalDragModeEnabled()) {
-      this.disableGlobalDragMode();
-    } else {
-      this.enableGlobalDragMode();
-    }
-  },
-  layerAddHandler({ layer }) {
-    // is this layer handled by leaflet-geoman?
-    const isRelevant = !!layer.pm && !layer._pmTempLayer;
-
-    // do nothing if layer is not handled by leaflet so it doesn't fire unnecessarily
-    if (!isRelevant) {
-      return;
-    }
-
-    // re-enable global removal mode if it's enabled already
-    if (this.globalRemovalEnabled()) {
-      this.disableGlobalRemovalMode();
-      this.enableGlobalRemovalMode();
-    }
-
-    // re-enable global edit mode if it's enabled already
-    if (this.globalEditEnabled()) {
-      this.disableGlobalEditMode();
-      this.enableGlobalEditMode();
-    }
-
-    // re-enable global drag mode if it's enabled already
-    if (this.globalDragModeEnabled()) {
-      this.disableGlobalDragMode();
-      this.enableGlobalDragMode();
-    }
-  },
-  disableGlobalRemovalMode() {
-    this._globalRemovalMode = false;
-    this.map.eachLayer(layer => {
-      layer.off('click', this.removeLayer, this);
-    });
-
-    // remove map handler
-    this.map.off('layeradd', this.layerAddHandler, this);
-
-    // toogle the button in the toolbar if this is called programatically
-    this.Toolbar.toggleButton('deleteLayer', this._globalRemovalMode);
-
-    this._fireRemovalModeEvent(false);
-  },
-  enableGlobalRemovalMode() {
-    const isRelevant = layer =>
-      layer.pm &&
-      !(layer.pm.options && layer.pm.options.preventMarkerRemoval) &&
-      !(layer instanceof L.LayerGroup);
-
-    this._globalRemovalMode = true;
-    // handle existing layers
-    this.map.eachLayer(layer => {
-      if (isRelevant(layer)) {
-        layer.on('click', this.removeLayer, this);
-      }
-    });
-
-    // handle layers that are added while in removal  xmode
-    this.map.on('layeradd', this.layerAddHandler, this);
-
-    // toogle the button in the toolbar if this is called programatically
-    this.Toolbar.toggleButton('deleteLayer', this._globalRemovalMode);
-
-    this._fireRemovalModeEvent(true);
-  },
-  _fireRemovalModeEvent(enabled) {
-    this.map.fire('pm:globalremovalmodetoggled', {
-      enabled,
-      map: this.map,
-    });
-  },
-  toggleGlobalRemovalMode() {
-    // toggle global edit mode
-    if (this.globalRemovalEnabled()) {
-      this.disableGlobalRemovalMode();
-    } else {
-      this.enableGlobalRemovalMode();
-    }
-  },
-  globalRemovalEnabled() {
-    return !!this._globalRemovalMode;
-  },
   getGlobalOptions() {
     return this.globalOptions;
   },
